Moved logging functionality to its own class. Framework delegates logging to KitCat::Logging instance
